requirements.txt
boto3
streamlit
pandas
pillow
awscli
numpy